refactor(es6): use es6 and fix #2
