+++ .npmignore
@@ -1,9 +1,5 @@
 .DS_Store
 node_modules
-dist
 coverage
 .idea
 .jshintrc
-/src/
-/examples/
-webpack.config.js
