+++ package.json
@@ -2,19 +2,15 @@
   "name": "vue-infinite-scroll",
   "version": "0.1.3",
   "description": "An infinite scroll directive for vue.js.",
-  "main": "lib/index.js",
+  "main": "vue-infinite-scroll.js",
+  "jsnext:main": "./src/index.js",
   "scripts": {
-    "test": "test",
-    "prepublish": "node_modules/babel-cli/bin/babel.js src --out-dir lib"
+    "test": "npm run build && echo lol",
+    "build": "rollup -c"
   },
-  "repository" : {
-    "type" : "git",
-    "url" : "https://github.com/ElemeFE/vue-infinite-scroll.git"
-  },
-  "babel": {
-    "presets": [
-      "es2015"
-    ]
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/ElemeFE/vue-infinite-scroll.git"
   },
   "keywords": [
     "infinite-scroll",
@@ -26,10 +22,8 @@
     "vue": "^1.0.10"
   },
   "devDependencies": {
-    "babel-cli": "^6.3.15",
-    "babel-core": "^6.3.15",
-    "babel-loader": "^6.2.0",
-    "babel-preset-es2015": "^6.3.13",
-    "webpack": "^1.12.9"
+    "babel-preset-es2015-rollup": "^1.1.1",
+    "rollup": "^0.25.4",
+    "rollup-plugin-babel": "^2.4.0"
   }
 }
